+++ server/src/main/java/org/opensearch/search/backpressure/settings/SearchShardTaskSettings.java
@@ -8,10 +8,15 @@
 
 package org.opensearch.search.backpressure.settings;
 
+import org.opensearch.ExceptionsHelper;
 import org.opensearch.common.settings.ClusterSettings;
 import org.opensearch.common.settings.Setting;
 import org.opensearch.common.settings.Settings;
 import org.opensearch.monitor.jvm.JvmStats;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.concurrent.TimeUnit;
 
 /**
  * Defines the settings related to the cancellation of SearchShardTasks.
@@ -22,21 +27,22 @@ public class SearchShardTaskSettings {
     private static final long HEAP_SIZE_BYTES = JvmStats.jvmStats().getMem().getHeapMax().getBytes();
 
     private static class Defaults {
-        private static final double TOTAL_HEAP_THRESHOLD = 0.05;
-        private static final double HEAP_THRESHOLD = 0.005;
+        private static final double TOTAL_HEAP_PERCENT_THRESHOLD = 0.05;
+        private static final double HEAP_PERCENT_THRESHOLD = 0.005;
         private static final double HEAP_VARIANCE_THRESHOLD = 2.0;
-        private static final long CPU_TIME_THRESHOLD = 15;
-        private static final long ELAPSED_TIME_THRESHOLD = 30000;
+        private static final int HEAP_MOVING_AVERAGE_WINDOW_SIZE = 100;
+        private static final long CPU_TIME_MILLIS_THRESHOLD = 15;
+        private static final long ELAPSED_TIME_MILLIS_THRESHOLD = 30000;
     }
 
     /**
      * Defines the heap usage threshold (in percentage) for the sum of heap usages across all search shard tasks
      * before in-flight cancellation is applied.
      */
-    private volatile double totalHeapThreshold;
-    public static final Setting<Double> SETTING_TOTAL_HEAP_THRESHOLD = Setting.doubleSetting(
-        "search_backpressure.search_shard_task.total_heap_threshold",
-        Defaults.TOTAL_HEAP_THRESHOLD,
+    private volatile double totalHeapPercentThreshold;
+    public static final Setting<Double> SETTING_TOTAL_HEAP_PERCENT_THRESHOLD = Setting.doubleSetting(
+        "search_backpressure.search_shard_task.total_heap_percent_threshold",
+        Defaults.TOTAL_HEAP_PERCENT_THRESHOLD,
         0.0,
         1.0,
         Setting.Property.Dynamic,
@@ -46,10 +52,10 @@ public class SearchShardTaskSettings {
     /**
      * Defines the heap usage threshold (in percentage) for an individual task before it is considered for cancellation.
      */
-    private volatile double heapThreshold;
-    public static final Setting<Double> SETTING_HEAP_THRESHOLD = Setting.doubleSetting(
-        "search_backpressure.search_shard_task.heap_threshold",
-        Defaults.HEAP_THRESHOLD,
+    private volatile double heapPercentThreshold;
+    public static final Setting<Double> SETTING_HEAP_PERCENT_THRESHOLD = Setting.doubleSetting(
+        "search_backpressure.search_shard_task.heap_percent_threshold",
+        Defaults.HEAP_PERCENT_THRESHOLD,
         0.0,
         1.0,
         Setting.Property.Dynamic,
@@ -70,12 +76,24 @@ public class SearchShardTaskSettings {
     );
 
     /**
+     * Defines the window size to calculate the moving average of heap usage of completed tasks.
+     */
+    private volatile int heapMovingAverageWindowSize;
+    public static final Setting<Integer> SETTING_HEAP_MOVING_AVERAGE_WINDOW_SIZE = Setting.intSetting(
+        "search_backpressure.search_shard_task.heap_moving_average_window_size",
+        Defaults.HEAP_MOVING_AVERAGE_WINDOW_SIZE,
+        0,
+        Setting.Property.Dynamic,
+        Setting.Property.NodeScope
+    );
+
+    /**
      * Defines the CPU usage threshold (in millis) for an individual task before it is considered for cancellation.
      */
-    private volatile long cpuTimeThreshold;
-    public static final Setting<Long> SETTING_CPU_TIME_THRESHOLD = Setting.longSetting(
-        "search_backpressure.search_shard_task.cpu_time_threshold",
-        Defaults.CPU_TIME_THRESHOLD,
+    private volatile long cpuTimeMillisThreshold;
+    public static final Setting<Long> SETTING_CPU_TIME_MILLIS_THRESHOLD = Setting.longSetting(
+        "search_backpressure.search_shard_task.cpu_time_millis_threshold",
+        Defaults.CPU_TIME_MILLIS_THRESHOLD,
         0,
         Setting.Property.Dynamic,
         Setting.Property.NodeScope
@@ -84,54 +102,70 @@ public class SearchShardTaskSettings {
     /**
      * Defines the elapsed time threshold (in millis) for an individual task before it is considered for cancellation.
      */
-    private volatile long elapsedTimeThreshold;
-    public static final Setting<Long> SETTING_ELAPSED_TIME_THRESHOLD = Setting.longSetting(
-        "search_backpressure.search_shard_task.elapsed_time_threshold",
-        Defaults.ELAPSED_TIME_THRESHOLD,
+    private volatile long elapsedTimeMillisThreshold;
+    public static final Setting<Long> SETTING_ELAPSED_TIME_MILLIS_THRESHOLD = Setting.longSetting(
+        "search_backpressure.search_shard_task.elapsed_time_millis_threshold",
+        Defaults.ELAPSED_TIME_MILLIS_THRESHOLD,
         0,
         Setting.Property.Dynamic,
         Setting.Property.NodeScope
     );
 
-    public SearchShardTaskSettings(Settings settings, ClusterSettings clusterSettings) {
-        totalHeapThreshold = SETTING_TOTAL_HEAP_THRESHOLD.get(settings);
-        clusterSettings.addSettingsUpdateConsumer(SETTING_TOTAL_HEAP_THRESHOLD, this::setTotalHeapThreshold);
+    /**
+     * Callback listeners.
+     */
+    public interface Listener {
+        void onHeapMovingAverageWindowSizeChanged(int newWindowSize);
+    }
 
-        heapThreshold = SETTING_HEAP_THRESHOLD.get(settings);
-        clusterSettings.addSettingsUpdateConsumer(SETTING_HEAP_THRESHOLD, this::setHeapThreshold);
+    private final List<Listener> listeners = new ArrayList<>();
+
+    public SearchShardTaskSettings(Settings settings, ClusterSettings clusterSettings) {
+        totalHeapPercentThreshold = SETTING_TOTAL_HEAP_PERCENT_THRESHOLD.get(settings);
+        clusterSettings.addSettingsUpdateConsumer(SETTING_TOTAL_HEAP_PERCENT_THRESHOLD, this::setTotalHeapPercentThreshold);
+
+        heapPercentThreshold = SETTING_HEAP_PERCENT_THRESHOLD.get(settings);
+        clusterSettings.addSettingsUpdateConsumer(SETTING_HEAP_PERCENT_THRESHOLD, this::setHeapPercentThreshold);
 
         heapVarianceThreshold = SETTING_HEAP_VARIANCE_THRESHOLD.get(settings);
         clusterSettings.addSettingsUpdateConsumer(SETTING_HEAP_VARIANCE_THRESHOLD, this::setHeapVarianceThreshold);
 
-        cpuTimeThreshold = SETTING_CPU_TIME_THRESHOLD.get(settings);
-        clusterSettings.addSettingsUpdateConsumer(SETTING_CPU_TIME_THRESHOLD, this::setCpuTimeThreshold);
+        heapMovingAverageWindowSize = SETTING_HEAP_MOVING_AVERAGE_WINDOW_SIZE.get(settings);
+        clusterSettings.addSettingsUpdateConsumer(SETTING_HEAP_MOVING_AVERAGE_WINDOW_SIZE, this::setHeapMovingAverageWindowSize);
 
-        elapsedTimeThreshold = SETTING_ELAPSED_TIME_THRESHOLD.get(settings);
-        clusterSettings.addSettingsUpdateConsumer(SETTING_ELAPSED_TIME_THRESHOLD, this::setElapsedTimeThreshold);
+        cpuTimeMillisThreshold = SETTING_CPU_TIME_MILLIS_THRESHOLD.get(settings);
+        clusterSettings.addSettingsUpdateConsumer(SETTING_CPU_TIME_MILLIS_THRESHOLD, this::setCpuTimeMillisThreshold);
+
+        elapsedTimeMillisThreshold = SETTING_ELAPSED_TIME_MILLIS_THRESHOLD.get(settings);
+        clusterSettings.addSettingsUpdateConsumer(SETTING_ELAPSED_TIME_MILLIS_THRESHOLD, this::setElapsedTimeMillisThreshold);
     }
 
-    public double getTotalHeapThreshold() {
-        return totalHeapThreshold;
+    public void addListener(Listener listener) {
+        listeners.add(listener);
     }
 
-    public long getTotalHeapThresholdBytes() {
-        return (long) (HEAP_SIZE_BYTES * getTotalHeapThreshold());
+    public double getTotalHeapPercentThreshold() {
+        return totalHeapPercentThreshold;
     }
 
-    private void setTotalHeapThreshold(double totalHeapThreshold) {
-        this.totalHeapThreshold = totalHeapThreshold;
+    public long getTotalHeapBytesThreshold() {
+        return (long) (HEAP_SIZE_BYTES * getTotalHeapPercentThreshold());
     }
 
-    public double getHeapThreshold() {
-        return heapThreshold;
+    private void setTotalHeapPercentThreshold(double totalHeapPercentThreshold) {
+        this.totalHeapPercentThreshold = totalHeapPercentThreshold;
     }
 
-    public long getHeapThresholdBytes() {
-        return (long) (HEAP_SIZE_BYTES * getHeapThreshold());
+    public double getHeapPercentThreshold() {
+        return heapPercentThreshold;
     }
 
-    private void setHeapThreshold(double heapThreshold) {
-        this.heapThreshold = heapThreshold;
+    public long getHeapBytesThreshold() {
+        return (long) (HEAP_SIZE_BYTES * getHeapPercentThreshold());
+    }
+
+    private void setHeapPercentThreshold(double heapPercentThreshold) {
+        this.heapPercentThreshold = heapPercentThreshold;
     }
 
     public double getHeapVarianceThreshold() {
@@ -142,19 +176,45 @@ public class SearchShardTaskSettings {
         this.heapVarianceThreshold = heapVarianceThreshold;
     }
 
-    public long getCpuTimeThreshold() {
-        return cpuTimeThreshold;
+    public int getHeapMovingAverageWindowSize() {
+        return heapMovingAverageWindowSize;
     }
 
-    private void setCpuTimeThreshold(long cpuTimeThreshold) {
-        this.cpuTimeThreshold = cpuTimeThreshold;
+    public void setHeapMovingAverageWindowSize(int heapMovingAverageWindowSize) {
+        this.heapMovingAverageWindowSize = heapMovingAverageWindowSize;
+
+        List<Exception> exceptions = new ArrayList<>();
+        for (Listener listener : listeners) {
+            try {
+                listener.onHeapMovingAverageWindowSizeChanged(heapMovingAverageWindowSize);
+            } catch (Exception e) {
+                exceptions.add(e);
+            }
+        }
+        ExceptionsHelper.maybeThrowRuntimeAndSuppress(exceptions);
     }
 
-    public long getElapsedTimeThreshold() {
-        return elapsedTimeThreshold;
+    public long getCpuTimeMillisThreshold() {
+        return cpuTimeMillisThreshold;
     }
 
-    private void setElapsedTimeThreshold(long elapsedTimeThreshold) {
-        this.elapsedTimeThreshold = elapsedTimeThreshold;
+    public long getCpuTimeNanosThreshold() {
+        return TimeUnit.MILLISECONDS.toNanos(getCpuTimeMillisThreshold());
+    }
+
+    private void setCpuTimeMillisThreshold(long cpuTimeMillisThreshold) {
+        this.cpuTimeMillisThreshold = cpuTimeMillisThreshold;
+    }
+
+    public long getElapsedTimeMillisThreshold() {
+        return elapsedTimeMillisThreshold;
+    }
+
+    public long getElapsedTimeNanosThreshold() {
+        return TimeUnit.MILLISECONDS.toNanos(getElapsedTimeMillisThreshold());
+    }
+
+    private void setElapsedTimeMillisThreshold(long elapsedTimeMillisThreshold) {
+        this.elapsedTimeMillisThreshold = elapsedTimeMillisThreshold;
     }
 }