+++ server/src/test/java/org/opensearch/search/backpressure/trackers/ElapsedTimeTrackerTests.java
@@ -1,0 +1,50 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ */
+
+package org.opensearch.search.backpressure.trackers;
+
+import org.opensearch.action.search.SearchShardTask;
+import org.opensearch.common.settings.ClusterSettings;
+import org.opensearch.common.settings.Settings;
+import org.opensearch.search.backpressure.settings.SearchBackpressureSettings;
+import org.opensearch.search.backpressure.settings.SearchShardTaskSettings;
+import org.opensearch.tasks.Task;
+import org.opensearch.tasks.TaskCancellation;
+import org.opensearch.test.OpenSearchTestCase;
+
+import java.util.Optional;
+
+import static org.opensearch.search.backpressure.SearchBackpressureTestHelpers.createMockTaskWithResourceStats;
+
+public class ElapsedTimeTrackerTests extends OpenSearchTestCase {
+
+    private static final SearchBackpressureSettings mockSettings = new SearchBackpressureSettings(
+        Settings.builder()
+            .put(SearchShardTaskSettings.SETTING_ELAPSED_TIME_MILLIS_THRESHOLD.getKey(), 100) // 100 ms
+            .build(),
+        new ClusterSettings(Settings.EMPTY, ClusterSettings.BUILT_IN_CLUSTER_SETTINGS)
+    );
+
+    public void testEligibleForCancellation() {
+        Task task = createMockTaskWithResourceStats(SearchShardTask.class, 1, 1, 0);
+        ElapsedTimeTracker tracker = new ElapsedTimeTracker(mockSettings, () -> 200000000);
+
+        Optional<TaskCancellation.Reason> reason = tracker.cancellationReason(task);
+        assertTrue(reason.isPresent());
+        assertEquals(1, reason.get().getCancellationScore());
+        assertEquals("elapsed time exceeded [200ms >= 100ms]", reason.get().getMessage());
+    }
+
+    public void testNotEligibleForCancellation() {
+        Task task = createMockTaskWithResourceStats(SearchShardTask.class, 1, 1, 150000000);
+        ElapsedTimeTracker tracker = new ElapsedTimeTracker(mockSettings, () -> 200000000);
+
+        Optional<TaskCancellation.Reason> reason = tracker.cancellationReason(task);
+        assertFalse(reason.isPresent());
+    }
+}